Rename: 폴더 위치 변경
Related to: #113
