+++ src/components/common/ChallengeItem/ChallengeItem.tsx
