+++ src/components/common/VerticalChallengeItems/VerticalChallengeItems.tsx
